requirements.txt
google-adk==0.0.1
google-genai==1.2.0
python-dotenv==1.2.1
requests==2.32.5
opentelemetry-api==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-requests==0.59b0
opentelemetry-instrumentation-urllib3==0.59b0
pydantic==2.5.0
pydantic-settings==2.2.1
httpx==0.25.2
aiofiles==23.2.1
uvicorn==0.24.0
fastapi==0.104.1
python-multipart==0.0.6
deprecated==1.3.1